+++ src/message_interface/message_handler.rs
@@ -36,6 +36,7 @@ use crate::{
         operations::transaction::{
             high_level::{create_alias::AliasOutputOptions, minting::mint_native_token::MintTokenTransactionDto},
             prepare_output::OutputOptions,
+            TransactionOptions,
         },
         types::{AccountBalanceDto, AccountIdentifier, TransactionDto},
         OutputDataDto,
@@ -210,7 +211,7 @@ impl WalletMessageHandler {
             }),
             Message::SetClientOptions(options) => {
                 convert_async_panics(|| async {
-                    self.account_manager.set_client_options(*options.clone()).await?;
+                    self.account_manager.set_client_options(*options).await?;
                     Ok(Response::Ok(()))
                 })
                 .await
@@ -288,9 +289,7 @@ impl WalletMessageHandler {
             } => {
                 convert_async_panics(|| async {
                     let duration = interval_in_milliseconds.map(Duration::from_millis);
-                    self.account_manager
-                        .start_background_syncing(options.clone(), duration)
-                        .await?;
+                    self.account_manager.start_background_syncing(options, duration).await?;
                     Ok(Response::Ok(()))
                 })
                 .await
@@ -361,6 +360,14 @@ impl WalletMessageHandler {
                         .await?
                         .map(|(e, _)| e);
                     Ok(Response::ParticipationEvent(event))
+                })
+                .await
+            }
+            #[cfg(feature = "participation")]
+            Message::GetParticipationEventIds(event_type) => {
+                convert_async_panics(|| async {
+                    let event_ids = self.account_manager.get_participation_event_ids(event_type).await?;
+                    Ok(Response::ParticipationEventIds(event_ids))
                 })
                 .await
             }
@@ -520,7 +527,7 @@ impl WalletMessageHandler {
                         .burn_native_token(
                             TokenId::try_from(&token_id)?,
                             U256::try_from(&burn_amount).map_err(|_| DtoError::InvalidField("burn_amount"))?,
-                            options,
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
@@ -529,7 +536,12 @@ impl WalletMessageHandler {
             }
             AccountMethod::BurnNft { nft_id, options } => {
                 convert_async_panics(|| async {
-                    let transaction = account_handle.burn_nft(NftId::try_from(&nft_id)?, options).await?;
+                    let transaction = account_handle
+                        .burn_nft(
+                            NftId::try_from(&nft_id)?,
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
+                        )
+                        .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
                 })
                 .await
@@ -556,7 +568,10 @@ impl WalletMessageHandler {
                         .transpose()?;
 
                     let transaction = account_handle
-                        .create_alias_output(alias_output_options, options)
+                        .create_alias_output(
+                            alias_output_options,
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
+                        )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
                 })
@@ -565,7 +580,10 @@ impl WalletMessageHandler {
             AccountMethod::DestroyAlias { alias_id, options } => {
                 convert_async_panics(|| async {
                     let transaction = account_handle
-                        .destroy_alias(AliasId::try_from(&alias_id)?, options)
+                        .destroy_alias(
+                            AliasId::try_from(&alias_id)?,
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
+                        )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
                 })
@@ -573,13 +591,18 @@ impl WalletMessageHandler {
             }
             AccountMethod::DestroyFoundry { foundry_id, options } => {
                 convert_async_panics(|| async {
-                    let transaction = account_handle.destroy_foundry(foundry_id, options).await?;
+                    let transaction = account_handle
+                        .destroy_foundry(
+                            foundry_id,
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
+                        )
+                        .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
                 })
                 .await
             }
             AccountMethod::GenerateAddresses { amount, options } => {
-                let address = account_handle.generate_addresses(amount, options.clone()).await?;
+                let address = account_handle.generate_addresses(amount, options).await?;
                 Ok(Response::GeneratedAddress(address))
             }
             AccountMethod::GetOutputsWithAdditionalUnlockConditions { outputs_to_claim } => {
@@ -664,7 +687,7 @@ impl WalletMessageHandler {
                         .decrease_native_token_supply(
                             TokenId::try_from(&token_id)?,
                             U256::try_from(&melt_amount).map_err(|_| DtoError::InvalidField("melt_amount"))?,
-                            options,
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
@@ -689,7 +712,7 @@ impl WalletMessageHandler {
                             TokenId::try_from(&token_id)?,
                             U256::try_from(&mint_amount).map_err(|_| DtoError::InvalidField("mint_amount"))?,
                             increase_native_token_supply_options,
-                            options.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::MintTokenTransaction(MintTokenTransactionDto::from(
@@ -704,7 +727,10 @@ impl WalletMessageHandler {
             } => {
                 convert_async_panics(|| async {
                     let transaction = account_handle
-                        .mint_native_token(NativeTokenOptions::try_from(&native_token_options)?, options.clone())
+                        .mint_native_token(
+                            NativeTokenOptions::try_from(&native_token_options)?,
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
+                        )
                         .await?;
                     Ok(Response::MintTokenTransaction(MintTokenTransactionDto::from(
                         &transaction,
@@ -733,7 +759,7 @@ impl WalletMessageHandler {
                                 .iter()
                                 .map(NftOptions::try_from)
                                 .collect::<Result<Vec<NftOptions>>>()?,
-                            options.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
@@ -749,7 +775,13 @@ impl WalletMessageHandler {
             } => {
                 convert_async_panics(|| async {
                     let output = account_handle
-                        .prepare_output(OutputOptions::try_from(&options)?, transaction_options.clone())
+                        .prepare_output(
+                            OutputOptions::try_from(&options)?,
+                            transaction_options
+                                .as_ref()
+                                .map(TransactionOptions::try_from_dto)
+                                .transpose()?,
+                        )
                         .await?;
                     Ok(Response::Output(OutputDto::from(&output)))
                 })
@@ -766,7 +798,7 @@ impl WalletMessageHandler {
                                 .iter()
                                 .map(AddressWithAmount::try_from)
                                 .collect::<Result<Vec<AddressWithAmount>>>()?,
-                            options.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::PreparedTransaction(PreparedTransactionDataDto::from(&data)))
@@ -782,7 +814,7 @@ impl WalletMessageHandler {
                                 .iter()
                                 .map(|o| Ok(Output::try_from_dto(o, token_supply)?))
                                 .collect::<Result<Vec<Output>>>()?,
-                            options.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::PreparedTransaction(PreparedTransactionDataDto::from(&data)))
@@ -803,7 +835,7 @@ impl WalletMessageHandler {
                 .await
             }
             AccountMethod::SyncAccount { options } => Ok(Response::Balance(AccountBalanceDto::from(
-                &account_handle.sync(options.clone()).await?,
+                &account_handle.sync(options).await?,
             ))),
             AccountMethod::SendAmount {
                 addresses_with_amount,
@@ -816,7 +848,7 @@ impl WalletMessageHandler {
                                 .iter()
                                 .map(AddressWithAmount::try_from)
                                 .collect::<Result<Vec<AddressWithAmount>>>()?,
-                            options.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
@@ -834,7 +866,7 @@ impl WalletMessageHandler {
                                 .iter()
                                 .map(AddressWithMicroAmount::try_from)
                                 .collect::<Result<Vec<AddressWithMicroAmount>>>()?,
-                            options.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
@@ -847,7 +879,10 @@ impl WalletMessageHandler {
             } => {
                 convert_async_panics(|| async {
                     let transaction = account_handle
-                        .send_native_tokens(addresses_native_tokens.clone(), options.clone())
+                        .send_native_tokens(
+                            addresses_native_tokens.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
+                        )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
                 })
@@ -859,7 +894,10 @@ impl WalletMessageHandler {
             } => {
                 convert_async_panics(|| async {
                     let transaction = account_handle
-                        .send_nft(addresses_nft_ids.clone(), options.clone())
+                        .send_nft(
+                            addresses_nft_ids.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
+                        )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
                 })
@@ -881,7 +919,7 @@ impl WalletMessageHandler {
                                 .iter()
                                 .map(|o| Ok(Output::try_from_dto(o, token_supply)?))
                                 .collect::<crate::Result<Vec<Output>>>()?,
-                            options.clone(),
+                            options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
                     Ok(Response::SentTransaction(TransactionDto::from(&transaction)))